Fixes and updates. See description.
fixed js to assign a "required" text to fields marked as required.
Added iframe_check.js which is enqueued on all pages except login and
register. If detected it is running in an iframe it will redirect the
browser in the parent window.
Added functionality for letting the user choose a password at
registration.
Fixed redirect loop bug when checking if fields that are required are
missing from a users profile.